chore: move facility and hazard map folder
