+++ web-client/src/pages/Admin/FacilityManagement/components/FacilityIcon.tsx
