+++ web-client/src/pages/Admin/FacilityManagement/components/MapComponent.tsx
@@ -1,13 +1,13 @@
 import { memo } from "react";
 // Services / API
-import { useAppDispatch, useAppSelector } from "../../../store/hooks";
+import { useAppDispatch, useAppSelector } from "../../../../store/hooks";
 import {
   selectAddMode,
   selectTempMarkerPos,
   setTempMarkerPos,
-} from "../../../store/slices/facilitySlice";
+} from "../../../../store/slices/facilitySlice";
 // Google Maps API
-import { GOOGLE_MAP_API_KEY } from "../../../api.config";
+import { GOOGLE_MAP_API_KEY } from "../../../../api.config";
 import { GoogleMap, MarkerF, useJsApiLoader } from "@react-google-maps/api";
 
 const containerStyle = {
