+++ web-client/src/pages/Admin/HazardReportsManagement/components/HazardReportsList.tsx
@@ -1,5 +1,5 @@
 import { memo } from "react";
-import { THazardReport } from "../types/hazardReport";
+import { THazardReport } from "../../../../types/hazardReport";
 import HazardReportItem from "./HazardReportItem";
 
 type TProps = {
