+++ web-client/src/services/facilityQuery.ts
@@ -1,4 +1,4 @@
-import { TFacility } from "../pages/MapManagement/types/emergencyFacility";
+import { TFacility } from "../types/emergencyFacility";
 import { rootApi } from "./rootApi";
 
 export const facilityQueryApi = rootApi.injectEndpoints({
