+++ web-client/src/store/slices/facilitySlice.ts
@@ -1,6 +1,6 @@
 import { createSelector, createSlice, PayloadAction } from "@reduxjs/toolkit";
 
-import { TFacility } from "../../pages/MapManagement/types/emergencyFacility";
+import { TFacility } from "../../types/emergencyFacility";
 
 import { RootState } from "../store";
