+++ web-client/src/types/hazardReport.d.ts
@@ -1,4 +1,4 @@
-import { User } from "../../../types/user";
+import { User } from "./user";
 
 export type THazardReport = {
   category: string;
